Refactored the cryptography-service as crypto-service to be an injectable (inversify). Injected that into the jwt-service. Updated the tokenwall middleware to get the injectable jwt-service from the DI container. Updated the token.test.ts file to get the jwt-service from the container as well, but i'm not sure if that is the best way to do it, still it works for now. Trying to fix the GitHub workflow/action that is currently failing because of tests, but works fine locally.
