+++ .github/workflows/build.yml
@@ -10,8 +10,6 @@ jobs:
     steps:
     - name: Checkout repository
       uses: actions/checkout@v2
-      with:
-        fetch-depth: 2
 
     - name: Set up Node.js
       uses: actions/setup-node@v2
